refactor(lox-core)!: move FrequencyBand and FrequencyRange to new comms module
