+++ crates/lox-comms/src/terminal.rs
@@ -26,7 +26,7 @@ use crate::pointing::Pointing;
 use crate::receiver::Receiver;
 use crate::transmitter::AmplifierTransmitter;
 use crate::{LinkBudgetError, ROOM_TEMPERATURE};
-use lox_core::units::FrequencyRange;
+use lox_core::comms::FrequencyRange;
 
 /// A transmit chain: an antenna fed by a transmitter.
 ///
@@ -68,7 +68,7 @@ impl TxChain {
     /// Creates a new transmit chain.
     ///
     /// `band` accepts anything convertible into a [`FrequencyRange`],
-    /// including [`FrequencyBand`](lox_core::units::FrequencyBand) letter
+    /// including [`FrequencyBand`](lox_core::comms::FrequencyBand) letter
     /// codes. Rejects a non-finite or negative feed loss.
     pub fn new(
         antenna: impl Into<Antenna>,
@@ -174,7 +174,7 @@ impl RxChain {
     /// Creates a new receive chain.
     ///
     /// `band` accepts anything convertible into a [`FrequencyRange`],
-    /// including [`FrequencyBand`](lox_core::units::FrequencyBand) letter
+    /// including [`FrequencyBand`](lox_core::comms::FrequencyBand) letter
     /// codes. Rejects a non-finite or negative feed loss or antenna noise
     /// temperature.
     pub fn new(
@@ -739,7 +739,7 @@ mod tests {
 
     #[test]
     fn test_band_accepts_ieee_letter_bands() {
-        use lox_core::units::FrequencyBand;
+        use lox_core::comms::FrequencyBand;
 
         let tx = TxChain::new(
             ConstantAntenna::new(46.0.db()).unwrap(),
