+++ crates/lox-core/src/lib.rs
@@ -14,6 +14,7 @@
 extern crate alloc;
 
 pub mod anomalies;
+pub mod comms;
 pub mod coords;
 pub mod elements;
 pub mod f64;
